requirements_test.txt
# Test dependencies for Helios Energy Optimizer
pytest-homeassistant-custom-component
pytest-asyncio
